Renamed unistroke to strokes
